+++ dart/gui/experimental/detail/filament/application.cpp
@@ -292,6 +292,7 @@ int runFilamentGuiApplicationImpl(
           viewport,
           appOptions.scene,
           dartScene,
+          cameraController,
           selectionController,
           orbitLight,
           debugOverlays,